# [Cache Proxy] Use AC/CAS RPC server interfaces directly instead of invoking client interfaces via self RPC