--- a/enterprise/server/cmd/cache_proxy/cache_proxy.go
+++ b/enterprise/server/cmd/cache_proxy/cache_proxy.go
@@ -187,8 +187,6 @@ func startGRPCServers(env *real_environment.RealEnv) error {
 		return status.InternalErrorf("CacheProxy: error starting local bytestream gRPC server: %s", err.Error())
 	}
 	env.SetLocalByteStreamClient(bspb.NewByteStreamClient(conn))
-	env.SetLocalActionCacheClient(repb.NewActionCacheClient(conn))
-	env.SetLocalCASClient(repb.NewContentAddressableStorageClient(conn))
 
 	s, err := grpc_server.New(env, grpc_server.GRPCPort(), false, grpcServerConfig)
 	if err != nil {
@@ -260,12 +258,12 @@ func registerInternalGRPCServices(grpcServer *grpc.Server, env *real_environment
 	if err != nil {
 		return status.InternalErrorf("CacheProxy: error starting local contentaddressablestorage server: %s", err.Error())
 	}
-	repb.RegisterContentAddressableStorageServer(grpcServer, localCAS)
+	env.SetLocalCASClient(localCAS)
 
 	localAC, err := action_cache_server.NewActionCacheServer(env)
 	if err != nil {
 		return status.InternalErrorf("CacheProxy: error starting local actioncache server: %s", err.Error())
 	}
-	repb.RegisterActionCacheServer(grpcServer, localAC)
+	env.SetLocalActionCacheClient(localAC)
 	return nil
 }