--- a/server/environment/environment.go
+++ b/server/environment/environment.go
@@ -97,11 +97,11 @@ type Env interface {
 	GetSSLService() interfaces.SSLService
 	GetBuildEventServer() pepb.PublishBuildEventServer
 	GetGitHubStatusService() interfaces.GitHubStatusService
-	GetLocalCASClient() repb.ContentAddressableStorageClient
+	GetLocalCASServer() repb.ContentAddressableStorageServer
 	GetCASServer() repb.ContentAddressableStorageServer
 	GetLocalByteStreamClient() bspb.ByteStreamClient
 	GetByteStreamServer() bspb.ByteStreamServer
-	GetLocalActionCacheClient() repb.ActionCacheClient
+	GetLocalActionCacheServer() repb.ActionCacheServer
 	GetActionCacheServer() repb.ActionCacheServer
 	GetPushServer() rapb.PushServer
 	GetFetchServer() rapb.FetchServer