Benchmark results on my local machine (from `benchstat`):

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 7 5800X 8-Core Processor             
                    │   old.txt   │               new.txt               │
                    │   sec/op    │   sec/op     vs base                │
FindMissingBlobs-16   313.4µ ± 8%   301.1µ ± 9%        ~ (p=0.280 n=10)
BatchReadBlobs-16     377.9µ ± 4%   236.6µ ± 7%  -37.39% (p=0.000 n=10)
BatchUpdateBlobs-16   601.8µ ± 6%   461.1µ ± 6%  -23.38% (p=0.000 n=10)
GetTree-16            744.4µ ± 7%   256.3µ ± 6%  -65.56% (p=0.000 n=10)
geomean               479.9µ        302.9µ       -36.88%

                    │    old.txt    │               new.txt                │
                    │     B/op      │     B/op      vs base                │
FindMissingBlobs-16    46.50Ki ± 0%   46.52Ki ± 0%        ~ (p=0.590 n=10)
BatchReadBlobs-16      74.90Ki ± 0%   51.95Ki ± 0%  -30.64% (p=0.000 n=10)
BatchUpdateBlobs-16    127.6Ki ± 0%   104.2Ki ± 0%  -18.34% (p=0.000 n=10)
GetTree-16            122.01Ki ± 0%   61.37Ki ± 0%  -49.70% (p=0.000 n=10)
geomean                85.82Ki        62.70Ki       -26.94%

                    │   old.txt   │               new.txt               │
                    │  allocs/op  │  allocs/op   vs base                │
FindMissingBlobs-16    810.0 ± 0%    809.0 ± 0%        ~ (p=0.370 n=10)
BatchReadBlobs-16     1104.0 ± 0%    710.0 ± 0%  -35.69% (p=0.000 n=10)
BatchUpdateBlobs-16   1.866k ± 0%   1.468k ± 0%  -21.33% (p=0.000 n=10)
GetTree-16            2052.0 ± 0%    989.0 ± 0%  -51.80% (p=0.000 n=10)
geomean               1.360k         955.6       -29.75%
```

Not sure exactly how this will translate to prod load, but results look promising. Will do ByteStream next it's just more work because the server/client interfaces differ more significantly than for unary RPCs.

https://github.com/buildbuddy-io/buildbuddy-internal/issues/4782